refactor: ProductDetail extension 위치 수정 및 setPriceText 메서드 수정
- bargain price와 price가 같으면 price만 표시
